move to config file for client + server
