+++ app/entry.client.tsx
@@ -6,7 +6,7 @@ import { hydrateRoot } from 'react-dom/client'
 import { I18nextProvider, initReactI18next } from 'react-i18next'
 import { HydratedRouter } from 'react-router/dom'
 import { getInitialNamespaces } from 'remix-i18next/client'
-import i18nextOptions from './i18next-options'
+import { i18nextOptions } from './i18next-config'
 
 const hydrate = async () => {
 	await i18next
