--- app/i18next-options.ts
@@ -1,9 +1,0 @@
-import { type InitOptions } from 'i18next'
-
-export const supportedLanguages = ['en', 'de'] as const
-
-export default {
-	supportedLngs: supportedLanguages,
-	fallbackLng: 'en',
-	defaultNS: 'common',
-} satisfies InitOptions
