+++ app/middleware/i18next.ts
@@ -3,9 +3,8 @@ import { resolve } from 'node:path'
 import Backend from 'i18next-fs-backend/cjs' // Even though unintuitive, cjs is what we want: https://github.com/i18next/i18next-fs-backend/issues/57
 import { initReactI18next } from 'react-i18next'
 import { createI18nextMiddleware } from 'remix-i18next/middleware'
-import 'i18next'
 import { i18nCookie } from '~/cookies'
-import i18nextOptions from '~/i18next-options'
+import { i18nextOptions } from '~/i18next-config'
 
 const getNamespaces = () => {
 	return readdirSync(resolve(`./public/locales/${i18nextOptions.fallbackLng}/`))
